refactor: separate client and server
