+++ src/game/client_side/game_manager_c.h
@@ -3,10 +3,11 @@
 #include <functional>
 #include <thread>
 
-#include "../common/event_listener.h"
-#include "../common/event_loop.h"
-#include "../common/state_machine.h"
-#include "../ui/ui_manager.h"
+#include "../../common/event_listener.h"
+#include "../../common/event_loop.h"
+#include "../../common/state_machine.h"
+#include "../../ui/ui_manager.h"
+#include "state_c.h"
 
 namespace SailGame { namespace Game {
 
@@ -22,7 +23,7 @@ public:
 
 private:
     std::unique_ptr<EventLoop> mEventLoop;
-    std::unique_ptr<StateMachine> mStateMachine;
+    std::unique_ptr<StateMachine<State>> mStateMachine;
     std::unique_ptr<UIManager> mUIManager;
 
     std::unique_ptr<std::thread> mNetworkThread;
