+++ src/game/client_side/state_c.h
@@ -1,8 +1,8 @@
 #pragma once
 
-#include "card.h"
-#include "config.h"
-#include "../common/util.h"
+#include "../card.h"
+#include "../config.h"
+#include "../../common/util.h"
 
 namespace SailGame { namespace Game {
 
@@ -47,31 +47,10 @@ private:
     int mIndexOfNewlyDrawn{-1};
 };
 
-class Handcards {
-public:
-    void Draw(Card card);
-
-    void Draw(const std::vector<Card> &cards);
-
-    void Erase(int index);
-
-    int GetIndex(Card card) const;
-
-    int GetIndexOfNewlyDrawn(const Handcards &handcardsBeforeDraw) const;
-
-    bool Empty() const { return mCards.empty(); }
-
-    Card At(int index) const { return *std::next(mCards.begin(), index); }
-
-    int Number() const { return mCards.size(); }
-
-private:
-    std::multiset<Card> mCards;
-};
-
 struct State {
     GameState mGameState;
     std::vector<PlayerState> mPlayerStates;
     Handcards mHandcards;
 };
+
 }}
