+++ src/game/server_side/state_s.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "../card.h"
+
+namespace SailGame { namespace Game {
+
+struct State {
+    Deck mDeck;
+    DiscardPile mDiscardPile;
+    std::vector<Handcards> mHandcards;
+};
+
+}}
